20250915-1921

+++ getmodels.sh
@@ -1,7 +1,9 @@
 #!/usr/bin/env bash
 
 cd /home/user/Models
-git clone http://github.com/lstaalhagen/sim34313models
-cp sim34313models/Models*.zip .
-rm -rf sim34313models
+/usr/bin/git clone http://github.com/lstaalhagen/sim34313models
+if [ -d sim34313models ] ; then
+  cp sim34313models/Models*.zip .
+  rm -rf sim34313models
+fi
